Enhance ratios
Ratios can now be made (indirectly) in the way that you would expect:

```idris
test : Ratio
test = 2 / 3
```

This currently relies on a hole: I cannot figure out how to prove that,
if an integer is greater than zero, its natural-number representation is
non-zero.

One of the `Uninhabited` implementations in this change causes
compilation to be really slow.